Adding agent image registry-path to faq.md
In our cluster, we have a policy to disallow pulling of external images which are not explicitly allowed. That is why the agent doesn't work at first. For me, it was not so easy to find out the registry-path of the agent-image. That's why I want to add this at minimum to the FAQ.

Found out by running `kubectl get events`:
```
LAST SEEN   TYPE      REASON         OBJECT                         MESSAGE
4m31s       Warning   FailedCreate   job/mirrord-agent-nwraf7nolp   Error creating: admission webhook denied the request: [...] Container image ghcr.io/metalbear-co/mirrord:3.1.2 for container mirrord-agent has not been allowed.
```